requirements.txt
streamlit
pandas
google-generativeai
tabulate
pywebview
pyinstaller
streamlit-pdf-viewer
pdfplumber
Pillow